Rename TotalsFooterItem folder
