+++ assets/js/base/components/cart-checkout/totals/index.js
@@ -2,7 +2,7 @@ export { default as Subtotal } from './subtotal';
 export { default as TotalsCoupon } from './coupon';
 export { default as TotalsDiscount } from './discount';
 export { default as TotalsFees } from './fees';
-export { default as TotalsFooterItem } from './totals-footer-item';
+export { default as TotalsFooterItem } from './footer-item';
 export { default as TotalsItem } from './totals-item';
 export { default as TotalsShippingItem } from './totals-shipping-item';
 export { default as TotalsTaxesItem } from './totals-taxes-item';
